hr-copilot-back/pyproject.toml
[project]
name = "hr-copilot-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.3",
    "httpx>=0.28.1",
    "pyjwt>=2.8.0",
    "jwt>=1.4.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.43.0",
    "bcrypt==4.0.1",
]
